+++ Engine/Source/Thebe/EngineParts/TextureBuffer.cpp
@@ -48,7 +48,7 @@ TextureBuffer::TextureBuffer()
 		return false;
 	}
 	
-	UINT bytesPerPixel = this->GetBytesPerPixel();
+	UINT bytesPerPixel = (UINT)this->GetBytesPerPixel();
 	if (bytesPerPixel == 0)
 		return false;
 
@@ -203,8 +203,8 @@ TextureBuffer::TextureBuffer()
 		return false;
 
 	const D3D12_RESOURCE_DESC& resourceDesc = this->GetResourceDesc();
-	rootValue->SetValue("width", new JsonInt(resourceDesc.Width));
-	rootValue->SetValue("height", new JsonInt(resourceDesc.Height));
+	rootValue->SetValue("width", new JsonInt((long)resourceDesc.Width));
+	rootValue->SetValue("height", new JsonInt((long)resourceDesc.Height));
 	rootValue->SetValue("pixel_format", new JsonInt(resourceDesc.Format));
 	rootValue->SetValue("num_mips", new JsonInt(resourceDesc.MipLevels));
